requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-generativeai==0.3.2
python-dotenv==1.0.0
twilio==8.10.3
pandas==2.1.3